+++ R/tbl_uvregression.R
@@ -89,7 +89,7 @@
 tbl_uvregression <- function(data, method, y = NULL, x = NULL, method.args = NULL,
                              formula = "{y} ~ {x}",
                              exponentiate = FALSE, label = NULL,
-                             include = NULL, exclude = NULL,
+                             include = everything(), exclude = NULL,
                              hide_n = FALSE, show_single_row = NULL, conf.level = NULL,
                              estimate_fun = NULL, pvalue_fun = NULL, show_yesno = NULL,
                              tidy_fun = NULL) {